+++ src/lib/database.ts
@@ -56,6 +56,31 @@ async function updateDb() {
   );
 
   updateLine(`
+  CREATE TABLE \`leo_incidents\` (
+    \`id\` varchar(255) NOT NULL,
+    \`case_number\` int(11) NOT NULL,
+    \`officer_dept\` varchar(255) NOT NULL,
+    \`officer_name\` varchar(255) NOT NULL,
+    \`full_date\` text NOT NULL,
+    \`involved_officers\` text NOT NULL,
+    \`location\` varchar(255) NOT NULL,
+    \`officer_id\` varchar(255) NOT NULL,
+    \`narrative\` text NOT NULL,
+    \`injuries\` varchar(255) NOT NULL,
+    \`arrests_made\` varchar(255) NOT NULL,
+    \`firearms_involved\` varchar(255) NOT NULL,
+    PRIMARY KEY (\`id\`)
+  ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
+  `);
+
+  updateLine(`
+  CREATE TABLE \`seo_tags\` (
+    \`title\` varchar(255) DEFAULT 'SnailyCAD',
+    \`description\` text DEFAULT 'A free, fast, simple and secure open source CAD/MDT',
+    \`site_name\` varchar(255) DEFAULT NULL
+  ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
+  `);
+  updateLine(`
   CREATE TABLE \`seo_tags\` (
     \`title\` varchar(255) DEFAULT 'SnailyCAD',
     \`description\` text DEFAULT 'A free, fast, simple and secure open source CAD/MDT',