+++ src/state/items/officer/OfficerActions.ts
@@ -3,8 +3,17 @@ import { getErrorFromResponse, handleRequest, notify } from "@lib/utils";
 import { Dispatch } from "react";
 import { Officer } from "types/Officer";
 import { SocketEvents } from "types/Socket";
-import { IOfficer, Search, IOfficers, SearchNames, GetOfficerLogs } from "./OfficerTypes";
+import {
+  IOfficer,
+  Search,
+  IOfficers,
+  SearchNames,
+  GetOfficerLogs,
+  IIncidents,
+  GetAllOfficers,
+} from "./OfficerTypes";
 import lang from "src/language.json";
+import { OfficerIncident } from "types/OfficerIncident";
 
 export const weaponSearch = (serialNumber: string) => async (dispatch: Dispatch<Search>) => {
   try {
@@ -222,6 +231,60 @@ export const saveNote = (citizenId: string, note: string) => async (
     });
 
     return notify.success(lang.officers.added_note);
+  } catch (e) {
+    const error = getErrorFromResponse(e);
+    return notify.warn(error);
+  }
+};
+
+export const getAllOfficers = (headers?: any) => async (dispatch: Dispatch<GetAllOfficers>) => {
+  try {
+    const res = await handleRequest("/officer/all", "GET", {
+      cookie: headers?.cookie,
+      url: headers?.host,
+    });
+
+    dispatch({
+      type: "GET_ALL_OFFICERS",
+      officers: res.data.officers,
+    });
+
+    return true;
+  } catch (e) {
+    return false;
+  }
+};
+
+export const getIncidents = (headers?: any) => async (dispatch: Dispatch<IIncidents>) => {
+  try {
+    const res = await handleRequest("/officer/incidents", "GET", {
+      cookie: headers?.cookie,
+      url: headers?.host,
+    });
+
+    dispatch({
+      type: "GET_INCIDENTS",
+      incidents: res.data.incidents,
+    });
+
+    return true;
+  } catch (e) {
+    return false;
+  }
+};
+
+export const createIncident = (data: Partial<OfficerIncident>) => async (
+  dispatch: Dispatch<IIncidents>,
+) => {
+  try {
+    const res = await handleRequest("/officer/incidents", "POST", data);
+
+    dispatch({
+      type: "CREATE_INCIDENT",
+      incidents: res.data.incidents,
+    });
+
+    return notify.success(lang.officers.created_incident);
   } catch (e) {
     const error = getErrorFromResponse(e);
     return notify.warn(error);