+++ src/state/items/officer/OfficerTypes.ts
@@ -1,4 +1,5 @@
 import { Officer, OfficerLog } from "types/Officer";
+import { OfficerIncident } from "types/OfficerIncident";
 
 export interface Search {
   type: "PLATE_SEARCH" | "NAME_SEARCH" | "WEAPON_SEARCH";
@@ -31,4 +32,21 @@ export interface SearchNames {
   names: Name[];
 }
 
-export type Actions = Search | IOfficer | IOfficers | SearchNames | GetOfficerLogs;
+export interface GetAllOfficers {
+  type: "GET_ALL_OFFICERS";
+  officers: Officer[];
+}
+
+export interface IIncidents {
+  type: "GET_INCIDENTS" | "CREATE_INCIDENT" | "UPDATE_INCIDENT";
+  incidents: OfficerIncident[];
+}
+
+export type Actions =
+  | Search
+  | IOfficer
+  | IOfficers
+  | SearchNames
+  | GetOfficerLogs
+  | IIncidents
+  | GetAllOfficers;